jumpスクリプトとobject placerスクリプトを分離
