+++ Assets/Resources/Player.prefab
@@ -201,6 +201,7 @@ GameObject:
   - component: {fileID: -5961268942961339687}
   - component: {fileID: 3166705076227694470}
   - component: {fileID: 4095289390666997066}
+  - component: {fileID: -7118825407410889630}
   m_Layer: 0
   m_Name: Player
   m_TagString: Player
@@ -338,6 +339,18 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   camera: {fileID: 61005458}
   clone: {fileID: 8176462837158943120, guid: 28b4a1681dcd93644bcbf0bcb2f4bf7f, type: 3}
+--- !u!114 &-7118825407410889630
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5877155554332136540}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: bf1c8087be9ecbb42b23d6834373accf, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   jumpPower: 5
   rb: {fileID: 3080623177550988141}
 --- !u!1 &5877155554541635786
