+++ Assets/Scripts/JumpController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bf1c8087be9ecbb42b23d6834373accf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
